+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.2 - Add mob encyclopedia
+* Mob encyclopedia will be added to player inventory when first join (configurable).
+* Right click encyclopedia with mobs to collect cards.
+* Reduce some spam logs.
+* Add total collectable card count in collection book's tooltip
+* Building related tweaks
 ## 0.1 - Initial version
 
 ---